+++ core/nuts/src/main/java/net/thevpc/nuts/NutsPath.java
@@ -105,6 +105,7 @@ public interface NutsPath extends NutsFormattable,NutsStreamMetadataAware {
     String getContentType();
 
     String getBaseName();
+    String getLongBaseName();
 
     String getLastExtension();